fix func template

+++ matmul.cu
@@ -71,7 +71,7 @@ void cudaCheck(cudaError_t error, const char *file, int line) {
 template <int TILE_K>
 void runTest(std::vector<uint8_t> current_test_ab,
              uint16_t current_test_c,
-             std::vector<uint32_t> &current_result);
+             std::vector<uint16_t> &current_result);
 
 
 //----------------------------------------------------------------------------//
